requirements.txt
python-dotenv
hurry.filesize
pyTelegramBotAPI
